+++ bonus/scene/in_parsing.h
@@ -6,7 +6,7 @@
 /*   By: dkim2 <dkim2@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/03 14:29:26 by dkim2             #+#    #+#             */
-/*   Updated: 2022/08/11 13:50:34 by dkim2            ###   ########.fr       */
+/*   Updated: 2022/08/16 17:53:47 by dkim2            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 # define IN_PARSING_H
 # include "scene.h"
 
-t_light	*create_light(t_vec3 pos, float b);
+t_light	*create_light(t_vec3 pos, double b);
 void	free_lightlst(t_light *light);
 int		case_light(t_scene *scene, char **single_scene);
 int		case_ambient(t_scene *scene, char **single_scene);
