+++ bonus/scene/light.c
@@ -6,7 +6,7 @@
 /*   By: dkim2 <dkim2@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/07 15:11:03 by dkim2             #+#    #+#             */
-/*   Updated: 2022/08/11 13:50:23 by dkim2            ###   ########.fr       */
+/*   Updated: 2022/08/16 17:57:23 by dkim2            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ static void	ft_addlst_back(t_light *light_list, t_light *node)
 	last->next->next = NULL;
 }
 
-t_light	*create_light(t_vec3 pos, float b)
+t_light	*create_light(t_vec3 pos, double b)
 {
 	t_light	*light;
 
@@ -59,7 +59,7 @@ int	case_light(t_scene *scene, char **single_scene)
 {
 	t_light	*light;
 	t_vec3	pos;
-	float	brightness;
+	double	brightness;
 
 	if (ft_strsetlen(single_scene) != 3 && ft_strsetlen(single_scene) != 4)
 		return (FALSE);
